src/app/extensions/tacton/pages/configure/tacton-text-input/tacton-text-input.component.ts
import { ChangeDetectionStrategy, Component } from '@angular/core';

import { TactonFacade } from '../../../facades/tacton.facade';
import { TactonConfigParameterComponent } from '../tacton-config-parameter/tacton-config-parameter.component';

@Component({
  selector: 'ish-tacton-text-input',
  templateUrl: './tacton-text-input.component.html',
  changeDetection: ChangeDetectionStrategy.OnPush,
})
export class TactonTextInputComponent extends TactonConfigParameterComponent {
  constructor(protected facade: TactonFacade) {
    super(facade);
  }
}
